+++ docs/CHANGELOG.md
@@ -4,11 +4,13 @@
 
 ## 2.0.3
 
+- New: admins can delete incidents
 - New: added new features to disable
 - New: added citizen images in /citizen page
 - New: added search input for /citizen page
 - Minor: fixed minor bugs
 - Minor: removed database timeouts
+- Minor: fixed small bug with status buttons
 
 ## 2.0.2